Extract active raw query execution
